Crate peruse

Peruse is a basic parser combinator library for Rust.

Currently the project is split into several modules * parsers - contains the main combinators * slice_parsers - a few simple primitive parsers for handling slices of items * string_parsers - a few parsers for handling strings

Parsers work by essentially building a heirarchy of structs that all implement the Parser and Combinator traits.

Modules

parsers

slice_parsers
Slice parsers are parsers designed to work with slices of items. These parsers work by consuming one or more elements off the beginning of the slice, converting them into an output type, and return the output value along with the rest of the slice.

string_parsers