Function peruse::parsers::one_of

pub fn one_of<T: Parser>(t: Vec<T>) -> OneOfParser<T>

Create a parser that attempts to use each of the given parsers until one succeeds. If all the given parses are literally the exact same type, they can be unboxed, otherwise you'll have to box them using the boxed function.

Examples

Here both parsers have the same structure, and can be used unboxed

let input = [2,3,4];
let p1 = lit(2);
let p2 = lit(3);
let parser = one_of(vec![p1, p2]);
parser.parse(&input);

These parsers have different structure, thus different types, so they need to be boxed to be used.

let input = [2, 3, 4];
let p1 = lit(2).then(lit(3)).map(|(a, b)| a * b);
let p2 = lit(4);
let parser = one_of(vec![boxed(p1), boxed(p2)]);
parser.parse(&input);