Trait peruse::parsers::ParserCombinator

pub trait ParserCombinator: Parser + Clone {
    fn then<P: Parser<I=Self::I>>(&self, p: P) -> ChainedParser<Self, P> { ... }
    fn then_r<P: ParserCombinator<I=Self::I>>(&self, p: P) -> MapParser<Self::I, ChainedParser<Self, P>, P> { ... }
    fn then_l<P: ParserCombinator<I=Self::I>>(&self, p: P) -> MapParser<Self::I, ChainedParser<Self, P>, Self::O> { ... }
    fn repeat(&self) -> RepeatParser<Self> { ... }
    fn map<T, F: 'static + Fn(Self::O) -> T>(&self, f: F) -> MapParser<Self::I, Self, T> { ... }
    fn or<P: Parser<I=Self::I, O=Self::O>>(&self, p: P) -> OrParser<Self, P> { ... }
}

Combinator methods for slice parsers. In most cases, these methods copy the caller into a higher-order parser

Provided Methods

fn then<P: Parser<I=Self::I>>(&self, p: P) -> ChainedParser<Self, P>

Chain this parser with another parser, creating new parser that returns a tuple of their results

fn then_r<P: ParserCombinator<I=Self::I>>(&self, p: P) -> MapParser<Self::I, ChainedParser<Self, P>, P>

Chain this parser with another parser, but toss the value from this parser

fn then_l<P: ParserCombinator<I=Self::I>>(&self, p: P) -> MapParser<Self::I, ChainedParser<Self, P>, Self::O>

Chain this parser with another parser, but toss the value from the other parser

fn repeat(&self) -> RepeatParser<Self>

Create a new parser that will repeat this parser until it returns an error

fn map<T, F: 'static + Fn(Self::O) -> T>(&self, f: F) -> MapParser<Self::I, Self, T>

Map the value of this parser

fn or<P: Parser<I=Self::I, O=Self::O>>(&self, p: P) -> OrParser<Self, P>

Create a disjunction with another parser. If this parser produces an error, the other parser will be used