Function peruse::slice_parsers::matcher

pub fn matcher<T: Clone, U, F: 'static + Fn(T) -> Option<U>>(f: F) -> MatchParser<T, U, F>

Create a parser that matches against the first element in the slice and returns either a Some of an output value, or None if the element does not match. If None is returned the parser will return an error.

Examples

let p1 = matcher(|i| if (i < 10) { Some(i) } else { None });
let input = [4, 40];
let (v, rest) = p1.parse(&input).unwrap(); //Ok

let (v2, rest2) = p1.parse(rest).unwrap(); //Err